# src/main/resources/openapi/sustainability-api.yaml
openapi: 3.1.0
info:
  title: Sustainability API
  description: APIs for water supply, electricity usage, and waste management.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local server

tags:
  - name: sustainability
    description: Sustainability REST API

paths:
  /sustainability/water-supply/{cityName}:
    get:
      summary: Get all water supply records by city
      operationId: getWaterSupplyByCity
      tags:
        - sustainability-controller
      parameters:
        - name: cityName
          in: path
          required: true
          description: the city name used to get water supply
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WaterSupply"
  /sustainability/waste/{cityName}:
    get:
      summary: Get all waste records by city
      operationId: getWasteByCity
      tags:
        - sustainability-controller
      parameters:
        - name: cityName
          in: path
          required: true
          description: the city name used to get waste
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Waste"
  /sustainability/waste:
    post:
      summary: Create waste record
      operationId: createWasteRecord
      tags:
        - sustainability-controller
      requestBody:
        description: Waste management payload
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Waste"
      responses:
        "201":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Waste"
  /sustainability/electricity/{cityName}:
    get:
      summary: Get all electricity usages by city
      operationId: getElectricityByCity
      tags:
        - sustainability-controller
      parameters:
        - name: cityName
          in: path
          required: true
          description: the city name used to get electricity usage
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Electricity"
  /sustainability/electricity/upload:
    post:
      summary: Upload CSV or Excel file to create electricity usage records
      description: Accepts a CSV or Excel file and processes electricity usage data.
      operationId: uploadElectricityUsage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV or Excel file containing electricity usage data
      responses:
        "201":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Electricity"
        "400":
          description: Bad request (invalid file format)
        "500":
          description: Internal server error

components:
  schemas:
    WaterSupply:
      type: object
      properties:
        id:
          type: string
        cityName:
          type: string
        supplyDate:
          type: string
          format: date
        waterVolume:
          type: number
    Waste:
      type: object
      properties:
        id:
          type: string
        cityName:
          type: string
        collectionDate:
          type: string
          format: date
        wasteType:
          type: string
        wasteVolume:
          type: number
    Electricity:
      type: object
      properties:
        id:
          type: string
        cityName:
          type: string
        usageDate:
          type: string
          format: date
        consumptionKwh:
          type: number